ffvidenc: Port to 0.11

+++ ext/ffmpeg/gstffmpegvidenc.h
@@ -42,14 +42,14 @@ struct _GstFFMpegVidEnc
   gboolean discont;
 
   /* cache */
-  gulong bitrate;
+  gint bitrate;
   gint me_method;
   gint gop_size;
-  gulong buffer_size;
-  gulong rtp_payload_size;
+  gint buffer_size;
+  gint rtp_payload_size;
 
   guint8 *working_buf;
-  gulong working_buf_size;
+  gsize working_buf_size;
 
   /* settings with some special handling */
   guint pass;
